Changes to update RegionSelector and TrigInDetAnalysis

See merge request !589

Former-commit-id: 093c285f97ceb19218fadebffbdca2091fd04bbc

+++ Trigger/TrigAnalysis/TrigInDetAnalysisUser/scripts/TIDArun.sh
@@ -4,17 +4,7 @@ pwd
 
 echo "args: $*"
 
-BASEDIR=/afs/cern.ch/user/s/sutt/public
-
-
-# copy basic web page information 
-
-for FILE in style.css index.php; do
-  echo "checking file $FILE"
-  if [ ! -e $FILE ]; then 
-    cp  $BASEDIR/TIDAWeb/TIDAV/$FILE .
-  fi
-done
+BASEDIR=/eos/atlas/atlascerngroupdisk/proj-sit/trigindet
 
 echo 
 echo "copy reference file and run comparitor"
@@ -24,39 +14,69 @@ BASEREFDIR=$BASEDIR/TrigInDetValidationReference
 
 if [ -e TrkNtuple-0000.root ]; then 
 
-    RELEASE=$(TIDAreader.exe -ro data*.root | grep release | awk '{print $2}' | head -1)
+    RELEASE=$(TIDAreader.exe -ro TrkNtuple-0000.root | grep release | awk '{print $2}' | head -1)
 
     REFDIR=$BASEREFDIR/share
+
+    echo "RELEASE: $RELEASE"
 
     if [ -d $BASEREFDIR/$RELEASE ]; then 
         REFDIR=$BASEREFDIR/$RELEASE
     fi
 
     echo "fetching reference files from $REFDIR"
-     
-    REFFILE=$REFDIR/$2
+        
+    EXPERT=$(echo $1 $2 | grep expert) 
 
-    if [ -e  $REFFILE ]; then 
+    # if expert timing histos
+    if [ "x$EXPERT" != "x" ]; then 
+      if [ "x$RTTJOBNAME" != "x" ]; then 
+        EXPERTREF="expert-monitoring-$(echo $RTTJOBNAME | sed 's|TrigInDetValidation_||g')-ref.root"
+        REFFILE=$REFDIR/expert/$EXPERTREF
+      else
+        REFFILE=$REFDIR/expert/$2
+      fi      
+    else
+      REFFILE=$REFDIR/$2
+    fi
+
+    NOREF=
+
+    if [ ! -e $REFFILE ]; then 
+
+      echo "no reference file $REFFILE - will process without"
+
+      NOREF=--noref
+
+    else
+
       echo "copying $REFFILE" 
 
       LOCALFILE=$(basename $REFFILE)
 
-      # don;t bother to copy it if it already exists in this directory
-      if [ ! -e $LOCALFILE ]; then 
-         cp $REFFILE .
-      fi
-
-      ls -l
-
-      echo
-      echo "running comparitor " `date`
-      echo
-
-      TIDAcomparitor.exe $*
-
-    else
-      echo "no reference file $REFFILE" 
+      # don't bother to copy it if it already exists in this directory
+      [ -e $LOCALFILE ] || cp $REFFILE .
+  
     fi
 
+    ls -l
+
+    for arg in $*; do 
+       if $(echo "$arg" | grep -q "TIDA.*.dat"); then get_files -data "$arg"; fi 
+       if $(echo "$arg" | grep -q "Test.*.dat"); then get_files -data "$arg"; fi 
+    done
+
+    if [ "x$EXPERT" == "x" ]; then 
+        get_files -data TIDAhistos-vtx.dat
+        echo -e "\nrunning comparitor " $(date) "\n"
+        TIDAcomparitor.exe $* $NOREF
+    else
+        echo -e "\nrunning cpucost " $(date) "\n"
+        TIDAcpucost.exe $* $NOREF
+    fi
+
+    echo "finished postprocessing"
+
+    ls -l
 
 fi
